Add shapefile uploader
Adds a UI component to allow for shapefile uploads. This creates
a Girder Item inside a User's Minerva folder (creating that folder
if necessary), and uploads all of the shapefile inputs to that Item.
After uploading is complete, the shapefile is processed to create
a geojson File, which is added to the Item.  The geojson File is
then displayed as an overlay on a basemap.