Working communication between Datasets/Layers/Map panels
At this point the dataset panel can add a dataset which will
upload it to Girder, this can be either a shapefile or geojson. If
the dataset is a shapefile, Gaia will be run on the server to process
and create geojson from the shapefile.  The datasets will appear in
the dataset list after uploading, from the dataset list the datasets
can be added to the Layers panel, and they will be displayed on top
of the Map panel as features.