Change from .on to .listenTo
It's better to have views connected to models via listenTo
rather than setting themselves as an event handler with on.
If setting themselves as an event handler with on, the view
won't be garbage collected when the time is right because there
is still a reference to the view from the model.