Add support for deleting a session from the UI
This change involved re-ordering of the imports, as GeoJs includes
a version of jquery in its minified external dependencies, and Girder
also includes a version of jQuery.  Girder adds a function to jQuery,
girderModol, but this was being overwritten by GeoJs' inclusion of
jQuery in the import order before this change.