Add contourJsonReader to geo.fileReaders namespace
Currently this is just a stubbed out version of jsonReader.  Eventually
- once this actually reads in contour json data and puts it on a layer -
we should talk about rolling this back into geojs.  For now by putting
it in the minerva folder we don't have to worry about tests, code
coverage, different jshint formats etc -  but can still leverage the
geojs fileReader API which is arguably the 'right' way to implement
this. If this gets rolled into GeoJS, or someone implements a better
version in GeoJS,  then we can switch over to that.

This is a temporary design decision that is a compromise between speed
 and correctness- it lets us use the GeoJS readFiles api without writing
 "minerva" code that manages rendering to GeoJS maps.