Abstract geoJson logic
This just renames some variables which are being tasked with more than
just holding geoJSON.  It also introduces a "renderableP" predicate that
returns whether or not a dataset can be rendered to the map.  See
comments for why current implementation is not so great and some
suggestions on how to refactor.