Fix rendering issue on analysis completion
This makes a rather large change in the SessionView file.  Previously
the minerva routing code at the bottom of SessionView bound session view
initialization to g:changed events on DataPanel,  AnalysisPanel and
LayerPanel updates.  This ment that calls to fetch() triggered the
creation of a NEW SessionView.  Because SessionView.render() created new
DataPanel, AnalysisPanel, LayerPanel and MapPanel views this ment a
fetch on any of these collections created a complete re-render of the
interface.

This change moves Panel creation into SessionView's initialize function
and changes the SessionView routing to only trigger the initialization
of the SessionView (via the g:navigateTo trigger) once.  Like this the
SessionView persists through fetches and does not re-create all Panel
Views on collection fetches.

Finally,  the error itself (The entire collection appears to disappear
except for the last item on fetch). This is due to the paging functionality
implemented in girder.Collection()  It is resolved by changing the call to
collection.fetch() on line 108  to collection.fetch({}, true).  This
resets the offset and requests the entire collection from the server
rather than the 'next page.'

Practically  If we have 6 datasets,  then the collection's internal
offset value is set to 6.  We add a 7th dataset through the romanesco
analysis and on fetch()  we display the next page. Because of the sort
order this is always the last item alphabetically. On page refresh we
get all 7 dataset items and the internal offset is set to 7.