Move wms tile creation from MapPanel to WMSDataset
This commit moved the specifcs of creating the tileUrl of the wms
dataset from the MapPanel to the WmsDatasetModel.  There could be
an argument for keeping this logic in the MapPanel, as it does
relate to setting up the visualization of the WMS layer, and hence
could be considered part of the view, but it also specifies state on the
WmsDataset and so could live in the model.  I moved it to the model
to keep a large block of WMS specific code out of the MapPanel as
that seem to provide a cleaner organization.  This commit also changed
the WMS layers within the MapPanel from being stored in their own JS
object and are instead stored in this.datasets, as this fixed the problem
where WMS datasets would throw an exception when they were deleted from
the DataPanel.  This commit also fixes an error where datasets that
didn't have legends would throw exceptions when they were deleted.