Fix stale source layers and duplicate layer add
This commit provides a fix for two separate but related issues.  The
first is that when a different WMS source was selected to display layers
from the original (first selected), the layers shown were from the wrong
(original) source.  In attempting a fix for that, I had removed the
instance WmsLayersListWidget from the SourcePanel and created a new
widget each time the layers were requested to display.  If the same
source was displayed twice to have layers added with this solution,
there were now two (or more) views listening on the same form
submission, resulting in multiple creations of the checked layers.  This
commit creates a single instance WmsLayersListWidget in the SourcePanel
and updates it whenever the requested source changes.  Also, in the jade
template wms-add-layers was changed from an id to a class as it related
to multiple inputs.