Reworking log streaming
There were a lot of inefficiencies that needed to be addressed.

Mainly, the log stream from Cumulus returned *everything* - and then a
stream of log output, this was causing a lot of work to be done on the
modal open. It's been rectified to only stream new log events, and we
take more care in when we decide to trigger changes to the logs.