Support a geojson time series.
If the datatype is marked a geojson but the file actually contains an array where the first entry contains 'time' (a moment parsable-time) and 'geojson'), then treat it as a geojson timeseries.

There are animation controls is the dataset_type is changed to 'geojson-timeseries', which currently requires going to the girder item for the data set and changing the metadata.

The dataset has various animation properties set on it.  I don't currently see these being saved.

A geojson-timeseries file is expected to be a json file with a single array, each entry of which is an object with keys 'geojson' (the geojson for the frame), 'time' (a timestamp that is parsable by the moment.js module), and optionally 'label' (a text string to show for the frame which defaults to the time value).

There are some animation controls which can be optionally shown which currently do nothing.  These are for grouping data by time, and would require some code to support them.

Currently, animations are approximately the speed that is requested.  Precision is not guaranteed, and slow frames will make the UI for the frame slider hard too use.  This could be improved by debouncing calls and aligning with renderAnimationFrame, but probably aren't worth doing unless the datasets we are using are not performant enough.