Refactor CI to use sudo: required to allow testing with GDAL 2.
Note that this changes the travis build quite a bit -- it is now run as a single build instead of a matrix (yet is faster).  More tests are actually executed, as the some additional standard tests are run.  We now use geojs's built in mock for WebGL.

Although coverage is checked, it does not get reported properly to cdash.  This is no different than before, but should be addressed in another PR.