Fix downloading database data.
Most of the internal calls to `downloadDataset` expect a decoded geojson
object, but, for databases, we returned a generator function.  For
non-database datasets, when we downloaded the data, we first decoded the
geojson to an object, then reencoded it to json as part of the endpoint
processing.  This fixes the internal use of the database dataset and
avoids the waste of reencoding non-database datasets.

Until we add test infrastructure for the database, you can manually
verify this by try to zoom to the bounds of a database dataset.